rename kimi cache capability module
