+++ .artifacts/execute/2026-04-06_15-36-36_provider-cache-capability-layer.md
@@ -1,0 +1,73 @@
+---
+title: "provider-cache-capability-layer execution log"
+link: "provider-cache-capability-layer-execute"
+type: debug_history
+ontological_relations:
+  - relates_to: [[provider-cache-capability-layer-plan]]
+tags: [execute, provider-cache-capability-layer]
+uuid: "8E6A3B05-C0D6-4293-95C6-ABD3B7D152EA"
+created_at: "2026-04-06T20:36:36Z"
+owner: "tuna"
+plan_path: ".artifacts/plan/2026-04-05_14-10-47_provider-cache-capability-layer/PLAN.md"
+start_commit: "ce4ec28"
+end_commit: "7a57ada"
+env: {target: "local", notes: "execute-phase run from clean worktree; .darkforest files referenced by AGENTS.md were not present in this checkout."}
+---
+
+## Pre-Flight Checks
+- Branch: `feature/cache-mvp`
+- Rollback commit: `7a57ada`
+- DoR satisfied: yes
+- Access/secrets: not required for offline implementation
+- Fixtures/data: ready (`.artifacts/research/2026-04-04_19-16-00_kimi-code-caching-findings.md`, `.artifacts/execute/2026-04-05_14-00-30_kimi-chat-completions-raw.log`)
+
+## Task Log
+
+### T001
+- Status: completed by verification
+- Evidence:
+  - `src/types/options.rs` exposes `CacheOptions { key }`, threads it through `BaseStreamOptions`, `SimpleStreamOptions`, and `OpenAICompletionsOptions`, and includes `cache_options_replace_session_id_in_public_options`
+  - `src/types/mod.rs` and `src/lib.rs` re-export `CacheOptions`
+  - Acceptance test passed: `cargo test cache_options_replace_session_id_in_public_options`
+
+### T002
+- Status: completed by verification
+- Evidence:
+  - `src/cache/mod.rs`, `src/cache/capability.rs`, `src/cache/request.rs`, and `src/cache/usage.rs` exist and implement the central cache capability layer
+  - Kimi capability provides endpoint override, model override, header injection, and `prompt_cache_key` body mutation
+  - Acceptance test passed: `cargo test kimi_cache_capability_returns_request_mutations`
+
+### T003
+- Status: completed by verification
+- Evidence:
+  - `src/providers/shared/openai_like_runtime.rs` applies cache request preparation
+  - `src/providers/shared/http.rs` merges model, cache, and request headers deterministically
+  - `src/providers/shared/stream_blocks.rs` routes cache token normalization through `src/cache/usage.rs`
+  - Acceptance test passed: `cargo test cache_usage_normalizer_reads_openai_cached_token_fields`
+
+### T004
+- Status: completed by verification
+- Evidence:
+  - `src/providers/kimi.rs` builds Kimi requests through the cache capability layer while preserving `Model<AnthropicMessages>`
+  - `src/models/kimi.rs` keeps public Kimi typed as `AnthropicMessages`
+  - `src/stream/mod.rs` dispatches Kimi Anthropic models to `stream_kimi_messages(...)`
+  - Acceptance test passed: `cargo test kimi_runtime_builds_chat_completions_request_with_cache_contract`
+
+### T005
+- Status: completed by verification
+- Evidence:
+  - `README.md`, `docs/providers/architecture.md`, `docs/providers/kimi.md`, and `examples/provider_probe.rs` contain the new cache contract language and Kimi transport notes
+  - Acceptance grep passed: `rg -n "prompt_cache_key|User-Agent: KimiCLI|cache options" README.md docs/providers/kimi.md docs/providers/architecture.md examples/provider_probe.rs`
+
+## Gate Results
+- Acceptance tests:
+  - `cargo test cache_options_replace_session_id_in_public_options` -> pass
+  - `cargo test kimi_cache_capability_returns_request_mutations` -> pass
+  - `cargo test cache_usage_normalizer_reads_openai_cached_token_fields` -> pass
+  - `cargo test kimi_runtime_builds_chat_completions_request_with_cache_contract` -> pass
+- Validation grep:
+  - `rg -n "prompt_cache_key|User-Agent: KimiCLI|cache options" README.md docs/providers/kimi.md docs/providers/architecture.md examples/provider_probe.rs` -> pass
+- Full harness:
+  - `make harness` -> pass
+- Notes:
+  - No source-code edits were required during this execute-phase run because the current branch already contained the plan deliverables; this run verified and documented the existing implementation.
