+++ src/cache/kimi_cache_capability.rs
