+++ src/cache/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod capability;
+pub(crate) mod kimi_cache_capability;
 pub(crate) mod request;
 pub(crate) mod usage;
